refactor: 增加脚本隔离
